Backend/requirements.txt
Flask==2.2.2
requests==2.28.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==3.0.3
python-dotenv==0.21.0
werkzeug==2.2.2
